[refactor] 방 목록 조회에 userprofile담기
