+++ build.gradle
@@ -47,6 +47,8 @@ dependencies {
     implementation 'io.jsonwebtoken:jjwt-impl:0.11.2'
     implementation 'io.jsonwebtoken:jjwt-jackson:0.11.2'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
+    implementation 'io.github.cdimascio:dotenv-java:3.0.0'
+
 
 
     // swagger
