+++ src/main/java/com/umc/yeogi_gal_lae/api/room/dto/response/RoomResponse.java
@@ -1,16 +1,15 @@
-package com.umc.yeogi_gal_lae.api.room.dto;
+package com.umc.yeogi_gal_lae.api.room.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.ToString;
+import java.util.List;
 
 @Getter
 @Builder
-@ToString
 @AllArgsConstructor
 public class RoomResponse {
     private Long roomId; // 방 ID
-    private String name; // 방 이름
-    private Long masterId;
+    private String roomName; // 방 이름
+    private List<SimpleRoomMemberResponse> members;
 }
