+++ src/main/java/com/umc/yeogi_gal_lae/global/config/WebConfig.java
@@ -1,6 +1,7 @@
 package com.umc.yeogi_gal_lae.global.config;
 
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
@@ -10,5 +11,7 @@ public class WebConfig implements WebMvcConfigurer {
     public void addViewControllers(ViewControllerRegistry registry) {
         registry.addRedirectViewController("/swagger", "/swagger-ui/index.html");
         registry.addRedirectViewController("/swagger/", "/swagger-ui/index.html");
+
     }
+
 }
